Makefile: obtain missing max number of CAIDs per slot
The VDR_MAXCAID defines now number of supported CAIDs for a CAM slot.
From VDR version 1.5.3 this value is 64. Earlier versions uses 16 CAIDs.
Without this commit the value was always set to 16. All CAIDs from all
channels are divided into CAM slots - so previously the CI adapter
created more CAM slots with less CAIDs. It should work even then but
there is still some problem in plugin code because it is always sending
requests to first slot only. This way a bunch of CAIDs (from the other
slots) was not presented in AOT_CA_PMT, leading to problems after
previous commit (2e84cb0):
Use CA descriptors for channel provided by vdr in AOT_CA_PMT.

Now the problems should be workaround-ed because the 64 CAIDs which are
in first slot should be enough.