Use CA descriptors for channel provided by vdr in AOT_CA_PMT
VDR is caching CA descriptors information for every channel. This
information is also collected during EPG scan mode. If vdr has CA
descriptors, then the plugin is now using it to provide to camd.socket.
This way the step for opening demux by dvbapi plugin and obtain PMT
data is omited and channel should ZAP faster. If we don't have this
information (mainly after VDR start) then the plugin is trying to obtain
PMT the old way.

Many thanks for Narog for figuring this way of handling this and
providing first working code.