added adapter_index to socket messages
The plugin is using a single global connection to camd.socket.
In this case it was unable to distinguish between adapters
when receiving CA_SET_PID and CA_SET_DESCR.
A conflict was occurring when a PID numbers was the same across
adapters (even if it uses different descrambler indexes).
Now the PIDs are assigned per adapter, avoiding simultaneous
channel decryption problems.

An adapter_index byte was introduced in OSCam rev9095.