Fix simultaneous decrypting of the same channel using different devices
This situation generally should not happen in VDR, but there are
some circumstances where it may occur:
The primary device is always preferred for live viewing.
So no matter what other devices are receiving a particular channel,
when one tune to it live it will be attempted to do so on the
primary device. It's mainly because with full featured cards
the primary device can display (FTA) channels without any further
action by VDR (i.e. without Transfer Mode). Should a recording be
started that requires the primary device (because all other devices
are busy receiving other transponders), it will take over the
primary device, and the live channel will select another device
that currently happens to receive the live transponder.

This commit fixes eg. the problem of broken recording when zap to
the same channel as the recorded one (in the above scenario).

The above problem should not affect streaming plugins at all
(at least when they don't try to obtain a device for LiveTV,
like latest versions of XVDR and VNSI plugins).