Add support for "MasterCamSlot" feature (VDR >= 2.3.3)
This patch for the VDR plugin 'dvbapi' makes use of the "MasterCamSlot"
feature introduced in VDR version 2.3.3. With this, CAMs that are basically
the same CAM, just use separate interfaces for each device, can all refer
to one "master", which is the only one that will be visible to the user.
VDR also no longer needs to check whether any of the device/CAM combinations
can decrypt a particular channel. It just checks once and if that CAM doesn't
decrypt, it knows that all the other ones (that are "slaves" to this "master")
won't be able to do so, either. This greatly speeds up channel switching on
systems with more than one CAM.

Here's what the patch does:
- removes unused "device" and "cafd" from SCCIAdapter
- removes unused "softcsa" and "fullts" from SCCIAdapter
- removes unused function Adapter() from SCCIAdapter
- removes unnecessary #includes <vdr/dvbdevice.h> and <vdr/dvbci.h>
- now only creating *one* SCCIAdapter, independent of any device
- removes cardIndex from SCCIAdapter (there is now only one)
- the one SCCIAdapter creates several SCCAMSlot objects (one per device)